Rename exposed components to make their internal names are as the same as exporting names. - google-ads-account-card/connected.js -> google-ads-account-card/connected-google-ads-account-card.js - google-mc-account-card/connected-card.js -> google-mc-account-card/connected-google-mc-account-card.js
Address https://github.com/woocommerce/google-listings-and-ads/pull/1374#discussion_r839466722
