+++ js/src/components/google-ads-account-card/index.js
@@ -1,2 +1,2 @@
 export { default } from './google-ads-account-card';
-export { default as ConnectedGoogleAdsAccountCard } from './connected';
+export { default as ConnectedGoogleAdsAccountCard } from './connected-google-ads-account-card';
