+++ js/src/components/google-mc-account-card/index.js
@@ -1,2 +1,2 @@
 export { default } from './google-mc-account-card';
-export { default as ConnectedGoogleMCAccountCard } from './connected-card';
+export { default as ConnectedGoogleMCAccountCard } from './connected-google-mc-account-card';
